+++ src/main/java/com/mia/bookstore/repository/AuthorRepository.java
@@ -4,11 +4,11 @@ import com.mia.bookstore.model.Author;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.data.jpa.repository.Query;
 
-public interface AuthorRepository extends JpaRepository<Author, Integer> {
+public interface AuthorRepository extends JpaRepository<Author, Long> {
 
     @Query("FROM author a WHERE a.authorName = :authorName")
      Author findByAuthorName(String authorName);
 
-    Author findByAuthorId(Integer id);
+    Author findByAuthorId(Long id);
 
 }
